[codex-rs] align PAT tests with review feedback [ci changed_files]
